+++ src/anim/ui/ui_bounded.cpp
@@ -1,0 +1,19 @@
+/***************************************************************
+ * Copyright (C) 2020
+ *    Reshetnyak Ivan
+ ***************************************************************/
+
+#include "pch.h"
+
+#include "ui_bounded.h"
+
+namespace spectral {
+namespace ui {
+
+
+bounded::bounded( const rect<float> &BoundBox ) : BoundBox(BoundBox) {
+}
+
+
+} // End of 'ui' namespace
+} // End of 'spectral' namespace